% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AquaticLifeHistory.R
\docType{package}
\name{AquaticLifeHistory}
\alias{AquaticLifeHistory}
\title{Introduction to AquaticLifeHistory}
\description{
Estimate aquatic species life history using robust techniques.
This package supports users undertaking two types of analysis: 1) Growth from
length-at-age data, and 2) maturity analyses for length and/or age data.

Maturity analyses are performed using generalised linear model approaches incorporating
either a binomial or quasibinomial distribution.

Growth modelling is performed using the multimodel approach presented by
Smart et al. (2016) "Multimodel approaches in shark and ray growth studies:
strengths, weaknesses and the future" <doi:10.1111/faf.12154>.
}
\references{
To cite the AquaticLifeHistory package in publications, type citation('AquaticLifeHistory').
}
\author{
Jonathan Smart
}
