% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{boot_data}
\alias{boot_data}
\title{boot_data}
\usage{
boot_data(.data, n = 100)
}
\arguments{
\item{.data}{Length at Age data}

\item{n}{number of bootstrap iterations}
}
\value{
a grouped_df produced by the group_by function in dplyr
}
\description{
boot_data
}
