% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AquaticLifeHistory.R
\docType{data}
\name{growth_data}
\alias{growth_data}
\title{Length-at-age data for blacktip sharks}
\format{
A data frame with 294 rows and 3 variables
}
\usage{
data(growth_data)
}
\description{
A dataset containing the length-at-age data for common blacktip sharks (\emph{Carcharhinus limbatus})
   from Indonesia. This data was published in Smart et al (2015).
}
\details{
\itemize{
  \item Age. Number of growth bands determined from vertebral analysis
  \item Length. Total Length in mm determined via back-calculation
  \item Sex. Females (F) or males (M)
}
}
\references{
Smart et al (2015)  Age and growth of the common
   blacktip shark Carcharhinus limbatus from Indonesia, incorporating an improved approach to
   comparing regional population growth rates. African Journal of Marine Science 37:177-188.
   \url{https://www.tandfonline.com/doi/abs/10.2989/1814232X.2015.1025428}
}
\keyword{datasets}
