% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AquaticLifeHistory.R
\docType{data}
\name{maturity_data}
\alias{maturity_data}
\title{Age and length-at-maturity data for silky sharks}
\format{A data frame with 284 rows and 4 variables}
\usage{
data(maturity_data)
}
\description{
A dataset containing the length-at-maturity and age-at-maturity data for female silky sharks
   (\emph{Carcharhinus falciformis}) from Papua New Guinea. This data was published in Grant et al (2018)
}
\details{
\itemize{
  \item Tag. Unique identifier for each Shark
  \item Age. Number of growth bands determined from vertebral analysis
  \item Length. Total Length in cm
  \item Maturity. Binary maturity status: immature = 0 and mature = 1
}
}
\references{
Grant et al (2018)
   Life history characteristics of the silky shark (\emph{Carcharhinus falciformis})
   from the central west Pacific. Marine and Freshwater Research 69:562-573
   \url{http://www.publish.csiro.au/mf/MF17163}
}
\keyword{datasets}
