% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AquaticLifeHistory.R
\docType{package}
\name{AquaticLifeHistory}
\alias{AquaticLifeHistory}
\alias{AquaticLifeHistory-package}
\title{AquaticLifeHistory}
\description{
A package to estimate life histories of aquatic species using robust techniques.
}
\author{
Jonathan Smart
}
